authorizing User

+++ app/controllers/evints_controller.rb
@@ -4,10 +4,12 @@ class EvintsController < ApplicationController
     @evint = Evint.new
     @event_options = Event.all
     @interest = Interest.find(params[:id])
+    authorize @evint
   end
 
   def create
     @evint = Evint.new(evints_params)
+    authorize @evint
     if @evint.save
       flash[:notice] = "Your new evint was saved."
       redirect_to :back
@@ -21,6 +23,7 @@ class EvintsController < ApplicationController
      @event = Event.find(params[:id])
      @interest = Interest.find(params[:id])
      @evint = Evint.find_by(event_id:@event.id, interest_id:@interest.id)
+     authorize @evint
     
      if @evint.destroy
        flash[:notice] = "Your evint has been removed."
